Working account and adds for user session

+++ src-main/com/finance/tracker/controller/RemoveAccountServlet.java
@@ -3,23 +3,30 @@ package com.finance.tracker.controller;
 import java.io.IOException;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import com.finance.tracker.model.dao.AccountDAO;
 
 @WebServlet("/account/remove")
-public class RemoveAccountServlet extends HttpServlet {
+public class RemoveAccountServlet extends BaseServlet {
 	private static final long serialVersionUID = 1L;
 
 	protected void doGet(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
+		if (!super.isAuthenticated(request)) {
+			response.sendRedirect("./login");
+			return;
+		}
 		response.sendRedirect("../account");
 	}
 
 	protected void doPost(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
+		if (!super.isAuthenticated(request)) {
+			response.sendRedirect("./login");
+			return;
+		}
 		int accountId = Integer.parseInt(request.getParameter("id"));
 		new AccountDAO().deleteAccount(accountId);
 		response.sendRedirect("../account");
